feat: Empty 컴포넌트 분리 및 알림 container 크기 고정
