+++ src/view/group/dashboard/Ranking/EmptyRanking.tsx
@@ -1,14 +1,10 @@
-import { IcnEmpty } from "@/asset/svg";
-import {
-  emptyGuideStyle,
-  emptyRankingWrapper,
-} from "@/view/group/dashboard/Ranking/index.css";
+import Empty from "@/shared/component/Empty";
+import { emptyRankingWrapper } from "@/view/group/dashboard/Ranking/index.css";
 
 const EmptyRanking = () => {
   return (
     <div className={emptyRankingWrapper}>
-      <IcnEmpty width={100} height={52} />
-      <p className={emptyGuideStyle}>아무도 문제를 풀지 않았어요.</p>
+      <Empty guideText="아무도 문제를 풀지 않았어요." />
     </div>
   );
 };
