+++ src/view/group/dashboard/Ranking/index.css.ts
@@ -139,17 +139,7 @@ export const changeStyle = recipe({
 });
 
 export const emptyRankingWrapper = style({
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "center",
-  alignItems: "center",
-  gap: "3rem",
-
-  width: "100%",
   padding: "5.3rem 0 5.2rem",
-});
-
-export const emptyGuideStyle = style({
-  color: theme.color.mg4,
-  ...theme.font.Head2_SB_18,
+  width: "100%",
+  height: "100%",
 });
